+++ cypress/all-tests/api-tests/petEntity/api-test.get-pet.js
@@ -1,6 +1,6 @@
 import Chance from 'chance'
 import {createPet, getPetById} from "../../../service/petService"
-import {DATA_OPTIONS, getPetRequestData} from "../../../util/requestsDataGenerator";
+import {DATA_OPTIONS, getPetRequestData} from "../../../utils/requestsDataGenerator";
 import {API_URL} from "../../../service/apiSettings";
 
 describe('Tests for Get pet endpoint', () => {
